Updated directory structure:
* Created web and api folders
