--- src/components/helperUtils.ts
@@ -1,2 +1,0 @@
-// CSS UTILS
-export const CONTAINER_HEIGHT = "calc(100vh - 64px)" // Height of viewport excluding navbar height
