--- src/components/login-form-components/styles.scss
@@ -1,5 +1,0 @@
-.login__form-container {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
