--- src/layouts/root/index.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Outlet } from 'react-router-dom';
-import Navbar from '../../components/navbar';
-
-export default function RooLayout() {
-    return (
-        <>
-            <Navbar />
-            <Outlet />
-        </>
-    );
-}
