--- src/styles.scss
@@ -1,3 +1,0 @@
-body{
-    margin:0;
-}
