+++ web/nx.json
