+++ web/src/components/helperUtils.ts
@@ -1,0 +1,2 @@
+// CSS UTILS
+export const NAVBAR_HEIGHT = "64px"
