+++ web/src/components/login-form-components/login-form-components.styles.ts
@@ -1,0 +1,21 @@
+import { SxProps, Theme } from "@mui/material";
+
+export const styles: Record<string, SxProps<Theme>> = {
+    mainContainer: {
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+        height: "100%",
+    },
+    boxContainer: {
+        width: "400px",
+    },
+    title: {
+        marginBottom: "1rem",
+    },
+    formContainer: {
+        display: "flex",
+        flexDirection: "column",
+        gap: "1rem",
+    },
+};
