+++ web/src/components/login-form-components/login-form/login-form.types.ts
@@ -1,0 +1,4 @@
+export interface LoginFormTypes {
+    username: string;
+    pin: string;
+}
