+++ web/src/components/navbar/index.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import AppBar from "@mui/material/AppBar";
+import Box from "@mui/material/Box";
+import Toolbar from "@mui/material/Toolbar";
+import Typography from "@mui/material/Typography";
+import Button from "@mui/material/Button";
+
+export default function ButtonAppBar() {
+    return (
+        <Box sx={{ flexGrow: 1 }}>
+            <AppBar position="absolute">
+                <Toolbar>
+                    <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
+                        Attendance
+                    </Typography>
+                </Toolbar>
+            </AppBar>
+        </Box>
+    );
+}
