+++ web/src/components/user-dashboard-overview/index.tsx
@@ -1,0 +1,12 @@
+import { Box } from "@mui/material";
+import WelcomePunchIn from "./welcome-punch-in";
+
+const UserDashboardOverview = () => {
+    return (
+        <Box sx={{ padding: "2rem" }}>
+            <WelcomePunchIn />
+        </Box>
+    );
+};
+
+export default UserDashboardOverview;
