+++ web/src/pages/first-login/index.tsx
@@ -1,0 +1,13 @@
+import { Box } from "@mui/material";
+import FirstLoginForm from "../../components/login-form-components/first-login-form";
+import PageWrapper from "../../components/ui/page-wrapper";
+
+const FirstLogin = () => {
+    return (
+        <PageWrapper>
+            <FirstLoginForm />
+        </PageWrapper>
+    );
+};
+
+export default FirstLogin;
