+++ web/src/pages/login/index.tsx
@@ -1,0 +1,10 @@
+import { Box, BoxProps } from "@mui/material";
+import LoginForm from "../../components/login-form-components/login-form/index";
+import { NAVBAR_HEIGHT } from "../../components/helperUtils";
+import PageWrapper from "../../components/ui/page-wrapper";
+
+const Login = () => {
+    return <LoginForm />;
+};
+
+export default Login;
