+++ web/src/pages/user-dashboard/index.tsx
@@ -1,0 +1,14 @@
+import { Box } from "@mui/material";
+import { NAVBAR_HEIGHT } from "../../components/helperUtils";
+import UserDashboardOverview from "../../components/user-dashboard-overview";
+import PageWrapper from "../../components/ui/page-wrapper";
+
+const UserDashboard = () => {
+    return (
+        <PageWrapper>
+            <UserDashboardOverview />
+        </PageWrapper>
+    );
+};
+
+export default UserDashboard;
