+++ web/src/styles.scss
@@ -1,0 +1,10 @@
+body{
+    margin:0;
+} 
+
+* {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+}
+
