+++ web/src/validations/validation.ts
@@ -1,0 +1,6 @@
+import * as Yup from "yup";
+
+export const loginFormSchema = Yup.object().shape({
+    username: Yup.string().required("Username is required"),
+    pin: Yup.string().min(4, "PIN must be atlease 4 characters").required("PIN is required"),
+});
